age in years are floats

+++ parser.py
@@ -45,7 +45,7 @@ def load_data(data_folder):
             obj['latest_date_received'] = datetime.datetime.strptime(obj['latest_date_received'], '%d-%b-%y')
             obj['initial_date_received'] = datetime.datetime.strptime(obj['initial_date_received'], '%d-%b-%y')
             obj['age'] = int(obj['age'])
-            obj['age_yrs'] = int(obj['age_yrs'])
+            obj['age_yrs'] = float(obj['age_yrs'])
             obj['drugs'] = []
             obj['reactions'] = []
             reports[id] = obj
